+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * :sparkles: Stringx adds splitToSequence, split, lineSequence, lines, chunkedSequence, windowedSequence
 * :white_check_mark: Perfect Mathx testing
 * :white_check_mark: Perfect Numberx testing
+* :fire: Sequence no longer inherits Iterable